api/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
langchain-community
langchain-core
openai
anthropic
pydantic
python-multipart
faiss-cpu
unstructured
pdf2image
tesseract
docx2txt
pypdf